DiamondSquareTerrain: Split in multiple files.
